+++ src/Control/Carrier/Empty/Maybe.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE DeriveFunctor, FlexibleInstances, MultiParamTypeClasses, TypeOperators, UndecidableInstances #-}
+{-# LANGUAGE FlexibleInstances, GeneralizedNewtypeDeriving, MultiParamTypeClasses, TypeOperators, UndecidableInstances #-}
 module Control.Carrier.Empty.Maybe
 ( -- * Empty effect
   module Control.Effect.Empty
@@ -10,64 +10,33 @@ module Control.Carrier.Empty.Maybe
 , run
 ) where
 
-import Control.Applicative (Alternative (..), liftA2)
 import Control.Carrier
 import Control.Effect.Empty
-import Control.Monad (MonadPlus (..))
 import qualified Control.Monad.Fail as Fail
 import Control.Monad.Fix
 import Control.Monad.IO.Class
 import Control.Monad.Trans.Class
+import Control.Monad.Trans.Maybe
 
 -- | Run an 'Empty' effect, returning 'Nothing' for empty computations, or 'Just' the result otherwise.
 --
 --   prop> run (runEmpty empty)    === Nothing
 --   prop> run (runEmpty (pure a)) === Just a
 runEmpty :: EmptyC m a -> m (Maybe a)
-runEmpty = runEmptyC
+runEmpty = runMaybeT . runEmptyC
+{-# INLINE runEmpty #-}
 
-newtype EmptyC m a = EmptyC { runEmptyC :: m (Maybe a) }
-  deriving (Functor)
+newtype EmptyC m a = EmptyC { runEmptyC :: MaybeT m a }
+  deriving (Applicative, Functor, Monad, MonadFix, MonadIO, MonadTrans)
 
-instance Applicative m => Applicative (EmptyC m) where
-  pure = EmptyC . pure . Just
-  {-# INLINE pure #-}
-  EmptyC f <*> EmptyC a = EmptyC (liftA2 (<*>) f a)
-  {-# INLINE (<*>) #-}
-
--- $
---   prop> run (runEmpty empty) === Nothing
-instance Applicative m => Alternative (EmptyC m) where
-  empty = EmptyC (pure Nothing)
-  {-# INLINE empty #-}
-  EmptyC a <|> EmptyC b = EmptyC (liftA2 (<|>) a b)
-  {-# INLINE (<|>) #-}
-
-instance Monad m => Monad (EmptyC m) where
-  EmptyC a >>= f = EmptyC (a >>= maybe (pure Nothing) (runEmptyC . f))
-  {-# INLINE (>>=) #-}
-
+-- | 'EmptyC' passes 'Fail.MonadFail' operations along to the underlying monad @m@, rather than interpreting it as a synonym for 'empty' à la 'MaybeT'.
 instance Fail.MonadFail m => Fail.MonadFail (EmptyC m) where
   fail = lift . Fail.fail
   {-# INLINE fail #-}
 
-instance MonadFix m => MonadFix (EmptyC m) where
-  mfix f = EmptyC (mfix (runEmpty . maybe (error "mfix (EmptyC): function returned failure") f))
-  {-# INLINE mfix #-}
-
-instance MonadIO m => MonadIO (EmptyC m) where
-  liftIO = lift . liftIO
-  {-# INLINE liftIO #-}
-
-instance (Alternative m, Monad m) => MonadPlus (EmptyC m)
-
-instance MonadTrans EmptyC where
-  lift = EmptyC . fmap Just
-  {-# INLINE lift #-}
-
 instance (Carrier sig m, Effect sig) => Carrier (Empty :+: sig) (EmptyC m) where
-  eff (L Empty) = EmptyC (pure Nothing)
-  eff (R other) = EmptyC (eff (handle (Just ()) (maybe (pure Nothing) runEmptyC) other))
+  eff (L Empty) = EmptyC (MaybeT (pure Nothing))
+  eff (R other) = EmptyC (MaybeT (eff (handle (Just ()) (maybe (pure Nothing) runEmpty) other)))
   {-# INLINE eff #-}